sttFlask/requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
Flask==2.2.2
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-speech==2.17.3
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==23.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.7.1
requests==2.28.2
rsa==4.9
setuptools==67.1.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.12.0
